requirements.txt
-i https://pypi.org/simple
anthropic==0.40.0
arxiv~=2.1.0
bertopic~=0.17.0
boto3==1.35.90
cohere==5.13.12
colorcet==3.0.1
datamapplot~=0.5.0
einops~=0.6.1
instructor==1.4.1
jupyter~=1.0.0
langchain-anthropic==0.1.15
langchain-cohere==0.1.5
langchain-community==0.2.1
langchain-core==0.2.3
langchain-groq==0.1.4
langchain-openai==0.1.8
langchain-postgres==0.0.9
langchain-text-splitters==0.2.0
langchain-together==0.0.2.post1
langchain==0.2.1
langchainhub==0.1.13
litellm==1.60.0
matplotlib==3.10.0
nltk~=3.8.1
numba==0.58.1
numpy==1.24.4
open-clip-torch~=2.20.0
openai==1.55.3
pandas==2.0.3
pdf2image==1.17.0
pgvector==0.2.5
plotly==5.18.0
psycopg2-binary==2.9.7
pydantic==2.9.1
PyMuPDF==1.22.5
pynndescent==0.5.11
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dotenv~=1.0.0
safetensors>=0.3.0
scikit-learn~=1.3.0
selenium~=4.16.0
sentence-transformers==2.7.0
setuptools~=57.5.0
streamlit-plotly-events==0.0.6
streamlit==1.42.2
tokencost>=0.1.12
torch~=2.4.0
torchdiffeq~=0.2.3
torchsde~=0.2.5
transformers~=4.42.2
voyageai~=0.3.2
webdriver_manager~=4.0.2